# perf(PG): add index for image risk score