--- a/proto/storage/image.proto
+++ b/proto/storage/image.proto
@@ -41,7 +41,7 @@ message Image {
 
   reserved 6; // was map<string,string> clusterns_scopes
   int64 priority = 11; // @gotags: search:"Image Risk Priority,hidden"
-  float risk_score = 12; // @gotags: search:"Image Risk Score,hidden"
+  float risk_score = 12; // @gotags: search:"Image Risk Score,hidden" sql:"index=btree"
   oneof set_top_cvss {
     float top_cvss = 13; // @gotags: search:"Image Top CVSS,store"
   }